+++ platform/CodatPlatform/Models/Operations/UpdateCompanyResponse.cs
@@ -40,6 +40,6 @@ namespace Codat.Platform.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }