+++ platform/RELEASES.md
@@ -97,3 +97,13 @@ Based on:
 - [csharp v3.0.0] platform
 ### Releases
 - [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Platform/3.0.0 - platform
+
+## 2023-11-20 01:25:35
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.1] platform
+### Releases
+- [NuGet v3.0.1] https://www.nuget.org/packages/Codat.Platform/3.0.1 - platform